feat/RSS-ECOMM-3_15: add personal hook
